% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterMSn}
\alias{filterMSn}
\title{Filter product ions to MRM}
\usage{
filterMSn(object, topX, type, cluster, a, b, ...)
}
\arguments{
\item{object}{metaMSn object}

\item{topX}{select top X product ions}

\item{type}{local/...}

\item{cluster}{}

\item{...}{NOT used}
}
\value{
Tibble of products and info (infoTibble)
}
\description{
Filter product ions to MRM
}
\examples{
NULL
}
