% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_isotopes}
\alias{find_isotopes}
\title{Find isotope proportion list}
\usage{
find_isotopes(i, mz_x, mz_tol, binMz = 0.5, mz_only = T, predictedIso)
}
\arguments{
\item{i}{The i-th product ion, i <= N}

\item{mz_x}{A MSn-matrix with 2 columns and N rows}

\item{mz_tol}{mz tolerance}

\item{predictedIso}{A list of j predicted isopote proportion tibble,
j is the j-th isotope, length(.) == isotope number}
}
\value{
A vector of length j, indicate whether the i-th product ion
        is a j-th isotope
}
\description{
Find isotope proportion list
}
\examples{
NULL
}
