% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_iso_percent}
\alias{get_iso_percent}
\title{Get isotope proportion range up/down}
\usage{
get_iso_percent(mz = mz_x, iso1 = 0, iso2 = 1, quantile_2 = c(0.05,
  0.95), cpdLib = "kegg", mzWin = 50)
}
\arguments{
\item{mz}{m/z vector to be calculated}

\item{iso1}{isotope up}

\item{iso2}{isotope down}

\item{quantile_2}{quantile range
databases for e.g. the database kegg}

\item{cpdLib}{the database; see compoundLibraries() for a list of supported
databases}

\item{mzWin}{the mass window size for grouping compounds;
see massWindowSizes(compoundLibrary = "kegg") for a list of supported}
}
\value{
A list of 2-column tibbles with length equal to the multiple iso arg
}
\description{
Get isotope proportion range up/down
}
\examples{
NULL
}
