% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mergeOverlap}
\alias{mergeOverlap}
\title{Merge overlapping transitions}
\usage{
mergeOverlap(traTbl, RTw = 2)
}
\arguments{
\item{traTbl}{transition tibble (from toSkyline() and subset by InChIKey)}

\item{RTw}{RT window}
}
\value{
merged traTbl
}
\description{
Merge overlapping transitions
}
\examples{
NULL
}
