% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mspParser}
\alias{mspParser}
\title{Parse .MSP file to list}
\usage{
mspParser(x = range_idx["1"], quantile_2 = c(0.05, 0.95),
  cpdLib = "kegg", iso1 = 0, iso2 = 1:3, lib_vct = lib,
  nbTol = 0.8, mz_tol = 0.01, mz_only = T, ...)
}
\arguments{
\item{x}{index of compound info. (list/vector/tibble)}

\item{quantile_2}{isotope proportion quantile}

\item{cpdLib}{library for calculate isotope proportion}

\item{lib_vct}{.MSP file lines in (vector), without blank lines}

\item{nbTol}{merge mz tolerance}

\item{mz_tol}{mz tolerance of isotope}

\item{mz_only}{ignore isotope proportion, use mz only}

\item{...}{not used}
}
\value{
list with compound info. vector and mz/intensity tibble
}
\description{
Parse .MSP file to list
}
\examples{
NULL
}
