% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toSkyline}
\alias{toSkyline}
\title{Write to skyline traTable}
\usage{
toSkyline(infoTibble, deltaMz)
}
\arguments{
\item{infoTibble}{DB infoTable from xMStoDB/filterMSn}

\item{deltaMz}{delta mz with precursor mz}
}
\value{
transition list table of skyline
}
\description{
Write to skyline traTable
}
\examples{
NULL
}
