% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toSkylineDB}
\alias{toSkylineDB}
\title{convert skylineTransition to DB.msp}
\usage{
toSkylineDB(X, File = "./skyline-db.msp", overWrite = T)
}
\arguments{
\item{X}{retruned by traTbl_join %>% toSkyline()}

\item{File}{file dir to save}
}
\description{
convert skylineTransition to DB.msp
}
\examples{
NULL
}
