% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xAdduct}
\alias{xAdduct}
\title{calculate adduct formula and mz}
\usage{
xAdduct(cls, adducts, formulas, polarities)
}
\arguments{
\item{cls}{parallel cluster object}

\item{adducts}{vector of adduct}

\item{formulas}{vector of formula}

\item{polarities}{vector of polarity}
}
\value{
tibble with 5 columns
}
\description{
calculate adduct formula and mz
}
\examples{
NULL
}
