% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xMStoDB}
\alias{xMStoDB}
\title{intersection of full-MS and Library}
\usage{
xMStoDB(MS, DB, tol, N)
}
\arguments{
\item{MS}{cross ms to db by ppm}

\item{DB}{db tibble}

\item{tol}{mz tolerance ppm}

\item{N}{segment size}
}
\value{
filtered db tibble
}
\description{
intersection of full-MS and Library
}
\examples{
NULL
}
